requirements.txt
aniso8601==9.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Cython==0.29.23
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Login==0.5.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
#fonttools==4.30.0
fuzzywuzzy==0.15.0
gensim==4.1.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.0.0
kiwisolver==1.3.2
lxml==4.8.0
MarkupSafe==2.1.0
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyahocorasick==1.4.4
pycparser==2.21
pymongo==4.0.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.5
python-Levenshtein==0.12.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
SQLAlchemy==1.4.32
textrank4zh==0.3
threadpoolctl==3.1.0
tushare==1.2.84
urllib3==1.26.8
websocket-client==0.57.0
Werkzeug==2.0.3
wordcloud==1.8.1
WTForms==3.0.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
